mvtec-vqi/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
gradio>=5.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
pyyaml>=6.0.0
tqdm>=4.66.0
kaggle>=1.5.16
requests>=2.31.0
Pillow>=9.5.0
python-multipart>=0.0.6
